LayoutTests/platform/mac-bigsur/imported/w3c/web-platform-tests/preload/preload-default-csp.sub-expected.txt

CONSOLE MESSAGE: Refused to load localhost:8801/preload/resources/dummy.js because it appears in neither the style-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/preload/resources/dummy.css because it appears in neither the style-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/preload/resources/square.png because it appears in neither the img-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/fonts/CanvasTest.ttf because it appears in neither the font-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/preload/resources/white.mp4 because it appears in neither the media-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/preload/resources/sound_5.oga because it appears in neither the media-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/preload/resources/foo.vtt because it appears in neither the media-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load localhost:8801/preload/resources/dummy.js because it appears in neither the style-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/preload/resources/dummy.css because it appears in neither the style-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/preload/resources/square.png because it appears in neither the img-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/fonts/CanvasTest.ttf because it appears in neither the font-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/preload/resources/white.mp4 because it appears in neither the media-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/preload/resources/sound_5.oga because it appears in neither the media-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load http://localhost:8800/preload/resources/foo.vtt because it appears in neither the media-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: <link rel=preload> must have a valid `as` value
CONSOLE MESSAGE: <link rel=preload> cannot have the empty string as `as` value

PASS Makes sure that preload requests respect CSP
